+++ hot_fair_utilities/__init__.py
@@ -1,6 +1,6 @@
 from .georeferencing import georeference
-from .inference import predict, evaluate
+from .inference import evaluate, predict
 from .postprocessing import polygonize, vectorize
 from .preprocessing import preprocess
-from .training import train
+from .training import ramp, yolo_v8_v1
 from .utils import bbox2tiles, tms2img
