+++ hot_fair_utilities/training/yolo_v8_v1/__init__.py
@@ -1,0 +1,1 @@
+from .train import train
